==> single_cell/pan_cancer/notebooks/02_feature_engineering/04_align_with_gdsc.ipynb <==
import pandas as pd


# Load GDSC drug response data
gdsc_df = pd.read_parquet("../../../../bulk_state_of_the_art/data/processed/gdsc_final_cleaned.parquet")
# Load aligned bulk and SC embeddings
df = pd.read_parquet("../../data/processed/pancancer_scran_norm_top8000_30_pcs.parquet")

print("✅ Loaded data:")
print("GDSC shape:", gdsc_df.shape)
print("Bulk embeddings:", df.shape)


# Preview columns
print("📌 GDSC columns:", gdsc_df.columns)

# Make sure it has the required fields
required_cols = {"SANGER_MODEL_ID", "DRUG_ID", "LN_IC50"}
assert required_cols.issubset(set(gdsc_df.columns)), "❌ GDSC dataset missing required columns!"

# Drop missing entries
gdsc_df = gdsc_df.dropna(subset=["SANGER_MODEL_ID", "DRUG_ID", "LN_IC50"])
gdsc_df["SANGER_MODEL_ID"] = gdsc_df["SANGER_MODEL_ID"].astype(str)

print("✅ Cleaned GDSC data:", gdsc_df.shape)


# Merge using SANGER_MODEL_ID
merged = gdsc_df.merge(df, on="SANGER_MODEL_ID", how="inner")

print("✅ Merged with bulk embeddings:")
print("Shape:", merged.shape)
print(merged.head(3))

# Save
merged.to_parquet("../../data/processed/gdsc_pancancer_scran_norm_top8000_30_pcs.parquet", index=False)
print("💾 Saved to: data/training_data_bulk.csv")
